try to make autojump more robust to concurrent uses over network filesystems
When your home is on a network filesytems, multiple shells can try
to write to the autojump store concurrently. The previous system was
a bit dangerous since the name of the temp file was always the same.
This should be better, since the different autojump instances should
all write to different temporaries, and only the atomic rename() should
be concurrent. Time will tell...