Vendorize argparse
This adds the argparse library [1] as autojump_argparse and imports it
via sys.path mangling in case argparse does not exist (Python 2.6 or
below without argparse installed).

This makes autojump effectively work with Python 2.6 again by default.

I have not verified license compatibility, but given the intention of
the (backport) project this is likely OK (it is licensed under the
Python license).

1: http://code.google.com/p/argparse/

Fixes issue #121.